Change rating type to int

+++ Domain/Entities/Review.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
     public sealed class Review
     {
         public int Id { get; set; }
-        public decimal Rating { get; set; }
+        public int Rating { get; set; }
         public string Comment { get; set; }
         public int HotelId { get; set; }
         public Hotel Hotel { get; set; }
